app/agents/core_agents/business_agent.py
import logging
from typing import Any, Dict, List
from app.agents.core_agents.agent_types import Agent
from app.models.agent.belief import Belief
from app.models.agent.desire import Desire
from app.models.agent.intention import Intention
from app.models.agent.goal import Goal
from app.tools.reasoning_engine import ReasoningEngine
from pydantic import Field

logger = logging.getLogger(__name__)

class BusinessAgent(Agent):
    business_id: str = Field(..., description="The business ID associated with the agent")
    business_type: str = Field(..., description="The type of business this agent represents")
    financial_data: Dict[str, Any] = Field(default_factory=dict, description="Financial data of the business")

    def __init__(self, **data):
        super().__init__(**data)
        self._init_business_beliefs()
        self._init_business_desires()

    def _init_business_beliefs(self):
        self.add_belief(Belief(id="business_id", content={"id": self.business_id}, description=f"Business ID: {self.business_id}", certainty=1.0))
        self.add_belief(Belief(id="business_type", content={"type": self.business_type}, description=f"Business Type: {self.business_type}", certainty=1.0))

    def _init_business_desires(self):
        self.add_desire(Desire(id="increase_profit", description="Increase business profit", priority=0.9))
        self.add_desire(Desire(id="expand_market", description="Expand market share", priority=0.8))

    async def reason(self, context: Dict[str, Any]) -> Dict[str, Any]:
        reasoning_result = await super().reason(context)
        
        # Business-specific reasoning logic
        if "financial_data" in context:
            self._update_financial_beliefs(context["financial_data"])
        
        return reasoning_result

    def _update_financial_beliefs(self, financial_data: Dict[str, Any]):
        for key, value in financial_data.items():
            self.add_belief(Belief(id=f"financial_{key}", content={key: value}, description=f"Financial data: {key}", certainty=0.9))

    async def act(self) -> None:
        """Perform the agent's main action cycle: reason, deliberate, and execute business actions."""
        try:
            context = self.get_current_state()
            await self.reason(context)
            self.deliberate()
            await self._execute_business_actions()
        except Exception as e:
            logger.error(f"Error during business agent action cycle: {str(e)}")

    def deliberate(self) -> None:
        """
        Deliberate on current beliefs and desires to form business-related intentions.
        """
        for desire in self.desires:
            if desire.id == "increase_profit" and self._should_focus_on_profit():
                self.add_intention(Intention(id="implement_cost_cutting", goal=Goal(id="reduce_costs", description="Reduce operational costs"), plan=None))
            elif desire.id == "expand_market" and self._should_focus_on_expansion():
                self.add_intention(Intention(id="launch_marketing_campaign", goal=Goal(id="increase_market_share", description="Increase market share"), plan=None))

    def _should_focus_on_profit(self) -> bool:
        # Logic to determine if the agent should focus on increasing profit
        return any(belief.content.get("profit_margin", 0) < 0.2 for belief in self.beliefs if "profit_margin" in belief.content)

    def _should_focus_on_expansion(self) -> bool:
        # Logic to determine if the agent should focus on market expansion
        return any(belief.content.get("market_share", 0) < 0.3 for belief in self.beliefs if "market_share" in belief.content)

    async def _execute_business_actions(self):
        for intention in self.intentions:
            if intention.id == "implement_cost_cutting":
                await self._implement_cost_cutting()
            elif intention.id == "launch_marketing_campaign":
                await self._launch_marketing_campaign()

    async def _implement_cost_cutting(self):
        # Implement cost-cutting measures
        print("Implementing cost-cutting measures")
        # In a real system, this would involve more complex logic and possibly interaction with other services

    async def _launch_marketing_campaign(self):
        # Launch a marketing campaign
        print("Launching marketing campaign")
        # In a real system, this would involve more complex logic and possibly interaction with other services